Support static mocking on Android 8
This extends support for static mocking back to Android 8. It's using
a very similar mechanism to the code for later version but has some
trickiness to deal with odd things in Android's code.

The method to attach TI agents is named differently in 8 (attachAgent
vs. attachJvmtiAgent) and is not publicly exposed. It also does not
support a customer classloader.

This last part makes for trouble: it relies on the system classloader
which only searches in system locations for shared libraries. It will,
however, accept an absolute pathname to a loadable library. But if
this pathname contains any equals signs, it'll get munged in this bit
of code:
https://cs.android.com/android/platform/superproject/+/android-8.0.0_r9:art/runtime/ti/agent.cc;l=130

All user-installed apps end up with equals signs in their unpacked
directory names (i.e. the place where the custom agent library is
unpacked to). This code works around that by creating a symbolic link
from its writable data directory to the shared object location.